requirements.txt
behave
asyncio
flake8
black
PyYAML
psutil
mkdocs
mkdocs-material
dotenv
paramiko
prettytable
colorlog